Add newly registered DOI
I had to complete registration one of a DOI outside the Collection. This is to updated our

rake data_cite:ingest_all

task about this fact. The DOI has been moved from allocated space to existing one.

I have also added gracefull handling of DataCite 404 response. Currently, the metadata for the DOI are not avaiable but should be avaiable by the time this task will be executed.